refactor(extractors): add common.py and deduplicate WordBox, clustering, text norm
- Add extractors/common.py with WordBox, RowCluster, normalize_text, compact_text,
  cluster_by_y, row_text, row_text_normalized, split_cluster_by_x_gap,
  DRAWING_NO_PATTERN, normalize_drawing_number_candidate
- raster_extractor: import from common, re-export for backward compat
- e055, e251: use common + raster.extract_words; remove local duplicates
- e142: import WordBox, cluster_by_y, normalize_text from common
- vector_extractor: use common for normalize_drawing_number_candidate

Made-with: Cursor
